Change timeout to support new semantics
Previously the timout sent from service to Pico meant "send me a
reply at this time". It now means "send me a reply by this time".

This change makes the response time from the Pico shorter by a
given leeway, so that the service doesn't timeout before the
message is received.